# cmd/registry/front/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Moonlight Box API",
    "description": "制品仓库管理系统 API 文档，支持 npm、maven、pypi、go、yum、apt 等多种包类型的代理、缓存和管理。",
    "version": "1.0.0"
  },
  "servers": [{"url": "/"}],
  "paths": {
    "/api/v1/ping": {"get": {"summary": "健康检查", "tags": ["System"], "responses": {"200": {"description": "成功"}}}},
    "/api/v1/auth/login": {"post": {"summary": "用户登录", "tags": ["Auth"], "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "required": ["username", "password"], "properties": {"username": {"type": "string"}, "password": {"type": "string"}}}}}}, "responses": {"200": {"description": "登录成功"}}}},
    "/api/v1/auth/logout": {"post": {"summary": "登出", "tags": ["Auth"], "security": [{"BearerAuth": []}], "responses": {"200": {"description": "登出成功"}}}},
    "/api/v1/auth/refresh": {"post": {"summary": "刷新 Token", "tags": ["Auth"], "requestBody": {"content": {"application/json": {"schema": {"type": "object", "required": ["refresh_token"], "properties": {"refresh_token": {"type": "string"}}}}}}, "responses": {"200": {"description": "刷新成功"}}}},
    "/api/v1/auth/profile": {"get": {"summary": "获取当前用户信息", "tags": ["Auth"], "security": [{"BearerAuth": []}], "responses": {"200": {"description": "成功"}}}, "put": {"summary": "更新用户信息", "tags": ["Auth"], "security": [{"BearerAuth": []}], "responses": {"200": {"description": "更新成功"}}}},
    "/api/v1/auth/password": {"put": {"summary": "修改密码", "tags": ["Auth"], "security": [{"BearerAuth": []}], "responses": {"200": {"description": "修改成功"}}}},
    "/api/v1/repositories": {"get": {"summary": "列出仓库", "tags": ["Repositories"], "security": [{"BearerAuth": []}], "responses": {"200": {"description": "成功"}}}, "post": {"summary": "创建仓库", "tags": ["Repositories"], "security": [{"BearerAuth": []}], "responses": {"201": {"description": "创建成功"}}}},
    "/api/v1/repositories/{name}": {"get": {"summary": "获取仓库详情", "tags": ["Repositories"], "security": [{"BearerAuth": []}], "parameters": [{"name": "name", "in": "path", "required": true, "schema": {"type": "string"}}], "responses": {"200": {"description": "成功"}}}, "put": {"summary": "更新仓库", "tags": ["Repositories"], "security": [{"BearerAuth": []}], "parameters": [{"name": "name", "in": "path", "required": true, "schema": {"type": "string"}}], "responses": {"200": {"description": "更新成功"}}}, "delete": {"summary": "删除仓库", "tags": ["Repositories"], "security": [{"BearerAuth": []}], "parameters": [{"name": "name", "in": "path", "required": true, "schema": {"type": "string"}}], "responses": {"200": {"description": "删除成功"}}}},
    "/api/v1/users": {"get": {"summary": "列出用户", "tags": ["Users"], "security": [{"BearerAuth": []}], "responses": {"200": {"description": "成功"}}}, "post": {"summary": "创建用户", "tags": ["Users"], "security": [{"BearerAuth": []}], "responses": {"201": {"description": "创建成功"}}}},
    "/api/v1/roles": {"get": {"summary": "列出角色", "tags": ["Roles"], "security": [{"BearerAuth": []}], "responses": {"200": {"description": "成功"}}}},
    "/api/v1/cache/caches": {"get": {"summary": "列出缓存", "tags": ["Cache"], "security": [{"BearerAuth": []}], "responses": {"200": {"description": "成功"}}}},
    "/api/v1/block-rules": {"get": {"summary": "列出阻断规则", "tags": ["BlockRules"], "security": [{"BearerAuth": []}], "responses": {"200": {"description": "成功"}}}, "post": {"summary": "创建阻断规则", "tags": ["BlockRules"], "security": [{"BearerAuth": []}], "responses": {"201": {"description": "创建成功"}}}},
    "/api/v1/backups": {"get": {"summary": "列出备份", "tags": ["Backups"], "security": [{"BearerAuth": []}], "responses": {"200": {"description": "成功"}}}, "post": {"summary": "创建备份", "tags": ["Backups"], "security": [{"BearerAuth": []}], "responses": {"201": {"description": "创建成功"}}}},
    "/api/v1/webhooks": {"get": {"summary": "列出 Webhooks", "tags": ["Webhooks"], "security": [{"BearerAuth": []}], "responses": {"200": {"description": "成功"}}}, "post": {"summary": "创建 Webhook", "tags": ["Webhooks"], "security": [{"BearerAuth": []}], "responses": {"201": {"description": "创建成功"}}}},
    "/api/v1/configs": {"get": {"summary": "获取系统配置", "tags": ["System"], "security": [{"BearerAuth": []}], "responses": {"200": {"description": "成功"}}}},
    "/api/v1/system/info": {"get": {"summary": "获取系统信息", "tags": ["System"], "security": [{"BearerAuth": []}], "responses": {"200": {"description": "成功"}}}},
    "/api/v1/dashboard/stats": {"get": {"summary": "获取仪表盘统计", "tags": ["Dashboard"], "security": [{"BearerAuth": []}], "responses": {"200": {"description": "成功"}}}},
    "/api/v1/audit/logs": {"get": {"summary": "获取审计日志", "tags": ["Audit"], "security": [{"BearerAuth": []}], "responses": {"200": {"description": "成功"}}}},
    "/api/v1/packages/search": {
      "get": {
        "summary": "搜索包",
        "description": "搜索包，支持关键字模糊搜索和精确匹配。所有参数均为可选。",
        "tags": ["Packages"],
        "parameters": [
          {"name": "q", "in": "query", "description": "关键字搜索（包名/描述模糊匹配）", "schema": {"type": "string"}},
          {"name": "type", "in": "query", "description": "包类型过滤（npm/maven/pypi/go/yum/apt等）", "schema": {"type": "string"}},
          {"name": "name", "in": "query", "description": "包名精确匹配", "schema": {"type": "string"}},
          {"name": "version", "in": "query", "description": "版本号精确查询", "schema": {"type": "string"}},
          {"name": "repository", "in": "query", "description": "按仓库名过滤", "schema": {"type": "string"}},
          {"name": "scope", "in": "query", "description": "搜索范围：name/description/all", "schema": {"type": "string", "default": "name"}},
          {"name": "sort", "in": "query", "description": "排序方式：downloads/name/updated_at", "schema": {"type": "string", "default": "downloads"}},
          {"name": "page", "in": "query", "description": "页码", "schema": {"type": "integer", "default": 1}},
          {"name": "page_size", "in": "query", "description": "每页数量（最大100）", "schema": {"type": "integer", "default": 20}}
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "list": {"type": "array", "items": {"$ref": "#/components/schemas/Package"}},
                    "total": {"type": "integer"},
                    "page": {"type": "integer"},
                    "page_size": {"type": "integer"},
                    "search_time_ms": {"type": "integer"}
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {"type": "http", "scheme": "bearer", "bearerFormat": "JWT", "description": "输入格式：Bearer {token}"}
    },
    "schemas": {
      "Package": {
        "type": "object",
        "properties": {
          "id": {"type": "integer"},
          "name": {"type": "string"},
          "display_name": {"type": "string"},
          "type": {"type": "string"},
          "description": {"type": "string"},
          "repository_id": {"type": "integer"},
          "repository_type": {"type": "string"},
          "repository_name": {"type": "string"},
          "homepage": {"type": "string"},
          "license": {"type": "string"},
          "download_count": {"type": "integer"},
          "versions": {"type": "array", "items": {"$ref": "#/components/schemas/PackageVersion"}}
        }
      },
      "PackageVersion": {
        "type": "object",
        "properties": {
          "id": {"type": "integer"},
          "package_id": {"type": "integer"},
          "version": {"type": "string"},
          "status": {"type": "string"},
          "published_at": {"type": "string", "format": "date-time"},
          "download_count": {"type": "integer"},
          "size_bytes": {"type": "integer"},
          "checksum_sha256": {"type": "string"},
          "checksum_md5": {"type": "string"}
        }
      }
    }
  }
}